# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "web-app-template API",
    "description": "REST API for web-app-template.",
    "version": "0.1.0"
  },
  "components": {
    "schemas": {
      "ApiErrorResponse": {
        "type": "object",
        "required": [
          "statusCode",
          "message",
          "error",
          "correlationId"
        ],
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "message": {
            "type": "string",
            "example": "Validation Error"
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          },
          "correlationId": {
            "type": "string",
            "example": "YevPQs"
          },
          "subErrors": {
            "type": "string",
            "description": "Optional list of sub-errors",
            "example": "incorrect email"
          }
        }
      },
...
      "CreateTaskRequest": {
        "type": "object",
        "required": [
          "lane"
        ],
        "properties": {
          "lane": {
            "type": "string",
            "example": "scrape",
            "description": "A lane from GET /lanes. An unknown one is a 400.",
            "minLength": 1
          },
          "params": {
            "type": "object",
            "additionalProperties": true,
            "description": "Parameters for the lane’s worker. Declared params are validated against their descriptors; undeclared keys ride along untouched."
          }
        },
        "title": "CreateTaskRequest"
      },
      "TaskStatus": {
        "anyOf": [
          {
            "type": "string",
            "enum": [
              "queued"
            ]
          },
          {
            "type": "string",
            "enum": [
              "running"
            ]
          },
          {
            "type": "string",
            "enum": [
              "ready"
            ]
          },
          {
            "type": "string",
            "enum": [
              "failed"
            ]
          },
          {
            "type": "string",
            "enum": [
              "cancelled"
            ]
          }
        ],
        "title": "TaskStatus",
        "description": "Where a task is in its lifecycle"
      },
...
    }
  },
  "paths": {
...
    "/api/v1/tasks": {
      "post": {
        "operationId": "createTask",
        "tags": [
          "tasks"
        ],
        "description": "Enqueue a job. Returns the task — with its handle — before any worker runs; the submit path never waits on the work itself.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "listTasks",
        "tags": [
          "tasks"
        ],
        "description": "Your tasks, newest first. Filter by status, lane and creation date; sort and page. Returns a bare array rather than the paginated envelope the rest of this API uses — the shape is fixed by the API contract.",
        "parameters": [
          {
            "schema": {
              "$ref": "#/components/schemas/TaskStatus"
            },
            "in": "query",
            "name": "status",
            "required": false,
            "description": "Where a task is in its lifecycle"
          },
          {
            "schema": {
              "type": "string",
              "example": "scrape"
            },
            "in": "query",
            "name": "lane",
            "required": false,
            "description": "Filter by lane"
          },
          {
            "schema": {
              "type": "string",
              "format": "date-time",
              "example": "2026-05-30T18:00:00.000Z"
            },
            "in": "query",
            "name": "from",
            "required": false,
            "description": "Only tasks created at or after this instant (ISO 8601)"
          },
          {
            "schema": {
              "type": "string",
              "format": "date-time",
              "example": "2026-05-30T18:00:00.000Z"
            },
            "in": "query",
            "name": "to",
            "required": false,
            "description": "Only tasks created at or before this instant (ISO 8601)"
          },
          {
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "enum": [
                    "asc"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "desc"
                  ]
                }
              ]
            },
            "in": "query",
            "name": "sort",
            "required": false,
            "description": "By creation time. Defaults to `desc` — newest first."
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "example": 20
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 0,
              "example": 0
            },
            "in": "query",
            "name": "offset",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/stats": {
      "get": {
        "operationId": "taskStats",
        "tags": [
          "tasks"
        ],
        "description": "Your task counts per status. Every status is present, zeroed when empty.",
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskStatsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/id/{id}": {
      "get": {
        "operationId": "getTaskById",
        "tags": [
          "tasks"
        ],
        "description": "One task by its immutable id. Handles are recycled, so this is the only address that keeps pointing at the same task once it retires.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "2cdc8ab1-6d50-49cc-ba14-54e4ac7ec231"
            },
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Task's id"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/id/{id}/history": {
      "get": {
        "operationId": "getTaskHistoryById",
        "tags": [
          "tasks"
        ],
        "description": "Every transition this task went through, oldest first. Addressed by id, so it keeps answering after the task retires and its handle is reused — the handle form resolves to whichever task holds the number now.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "2cdc8ab1-6d50-49cc-ba14-54e4ac7ec231"
            },
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Task's id"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskHistoryResponse"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{handle}": {
      "get": {
        "operationId": "getTask",
        "tags": [
          "tasks"
        ],
        "description": "One task by handle. Resolves to the active task holding that number.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "scrape-1"
            },
            "in": "path",
            "name": "handle",
            "required": true,
            "description": "The task handle, `<lane>-<n>`"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{handle}/result": {
      "get": {
        "operationId": "collectTaskResult",
        "tags": [
          "tasks"
        ],
        "description": "Take delivery of a finished task: returns the whole task with `result` populated and `collected: true`, and releases its handle number for reuse. Only a ready, not-yet-collected task can be collected — anything else is a 409.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "scrape-1"
            },
            "in": "path",
            "name": "handle",
            "required": true,
            "description": "The task handle, `<lane>-<n>`"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{handle}/history": {
      "get": {
        "operationId": "getTaskHistory",
        "tags": [
          "tasks"
        ],
        "description": "Every transition this task went through, oldest first, with timestamps. This is where a retry’s interim failure reason and a restart’s requeue are recorded. Resolves to whichever task holds the handle now; use the id form for a retired one.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "scrape-1"
            },
            "in": "path",
            "name": "handle",
            "required": true,
            "description": "The task handle, `<lane>-<n>`"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskHistoryResponse"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{handle}/cancel": {
      "post": {
        "operationId": "cancelTask",
        "tags": [
          "tasks"
        ],
        "description": "Cancel a queued or running task — a running worker is actually aborted, not left to finish quietly. Also dismisses a failed task, which is what releases its handle number. Any other state is a 409.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "scrape-1"
            },
            "in": "path",
            "name": "handle",
            "required": true,
            "description": "The task handle, `<lane>-<n>`"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/events": {
      "get": {
        "operationId": "streamEvents",
        "tags": [
          "tasks"
        ],
        "description": "Server-sent stream of lifecycle events for your tasks. Every frame carries an `id`, so a browser reconnects with `Last-Event-ID` — or curl with `?since=` — and receives exactly what it missed. `accepted`, `ready`, `failed` and `cancelled` are the contract; other types are informational and may be ignored.",
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 0,
              "example": 0
            },
            "in": "query",
            "name": "since",
            "required": false,
            "description": "Replay every event after this id before streaming live ones. A browser sends the `Last-Event-ID` header instead and does not need this; it is here so a curl client can replay too. The header wins when both are present."
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    }
  }
}